WIP cross platform porting.
-Link more necessary libraries in CMakeLists.
-Integrated cross platform memset/memcpy/move calls.
-Integrated cross platform system key definition.
-Process client events at once and carry on engine logic for better frame pace.
-static constexpr instead of static const for variables.
-Standardize std math usages.
-Avoid wide char hard-code for Linux.
-Removed more Win32 dependencies.
-Linux build is a success yay. Need to deal with runtime errors next.

+++ Runtime/Classes/Math.cpp
@@ -77,14 +77,14 @@ namespace UHMathHelpers
 
         if (M[0][0] == 1.0f)
         {
-            Yaw = std::atan2f(M[0][2], M[2][3]);
+            Yaw = std::atan2(M[0][2], M[2][3]);
             Pitch = 0;
             Roll = 0;
 
         }
         else if (M[0][0] == -1.0f)
         {
-            Yaw = std::atan2f(M[0][2], M[2][3]);
+            Yaw = std::atan2(M[0][2], M[2][3]);
             Pitch = 0;
             Roll = 0;
         }
@@ -164,7 +164,7 @@ namespace UHMathHelpers
 
     float RoundUpToMultiple(float InVal, float InMultiple)
     {
-        return std::roundf(InVal / InMultiple) * InMultiple;
+        return std::round(InVal / InMultiple) * InMultiple;
     }
 
     // matrix functions
@@ -530,7 +530,7 @@ void UHBoundingFrustum::Transform(UHBoundingFrustum& Out, UHMatrix4x4 M) const n
     float DotY = UHMathHelpers::UHVector3Dot(R1, R1);
     float DotZ = UHMathHelpers::UHVector3Dot(R2, R2);
     float MaxD = std::max(DotX, std::max(DotY, DotZ));
-    float Scale = std::sqrtf(MaxD);
+    float Scale = std::sqrt(MaxD);
 
     Out.Near = Near * Scale;
     Out.Far = Far * Scale;
